real_requirements.txt
# the following ones are required by the course
matplotlib
numpy
pandas
loguru
dvc
dvc[gc]
pillow
wandb
click
dvc==3.50.1
dvc[gs]
einops
ftfy
hydra-core
Jinja2
Kaggle==1.6.14
numpy
pre-commit
pytest
pytorch_lightning
snakeviz
torch_tb_profiler
prometheus-fastapi-instrumentator

# the following ones are required for our experiments
tensorboard
torch
torchvision
tqdm
transformers>=4.25.1
# xformers
albumentations
accelerate>=0.16.0
bitsandbytes
datasets
diffusers
kornia==0.6

fastapi
uvicorn

nvgpu

torchserve
torch-model-archiver
torch-workflow-archiver

# the following libraries are required to push the huggingface dataset to the hub
huggingface_hub

-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e git+https://github.com/openai/CLIP.git@main#egg=clip

prometheus-fastapi-instrumentator